requirements.txt
asgiref==3.8.1
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.2
Django==5.0.4
djangorestframework==3.15.1
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
uvicorn==0.29.0
zope.interface==6.3
python-dotenv==1.0.1
langchain-groq
websockets
celery
redis
django-celery-results
django-celery-beat
django-redis
langchain-community
neo4j
tiktoken
langchain
Pillow
boto3 
django-storages
dj_database_url
django-cors-headers
django-guardian
psycopg2-binary
whitenoise
django-six
six
pycryptodomex
stripe
djangorestframework-msgpack
msgpack-python
django-utils-six
webptools
icecream
pymongo
torch
transformers
sentence_transformers
einops
neomodel
django-neomodel
channels_redis
polars
scikit-image
orjson
PyMuPDF
llama_index_core
llama_index
qdrant_client
llama-index-vector-stores-qdrant
llama-index-embeddings-langchain
llama-index-llms-groq
FastEmbed
llama-index-readers-qdrant
llama-index-llms-langchain
jwt
pyautogen